+++ MapboxSearch/base/src/main/java/com/mapbox/search/base/utils/RequestTimeoutProvider.kt
@@ -1,0 +1,6 @@
+package com.mapbox.search.base.utils
+
+import androidx.annotation.RestrictTo
+
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP_PREFIX)
+fun defaultOnlineRequestTimeoutSeconds(): Int? = null