+++ MapboxSearch/offline/src/main/java/com/mapbox/search/offline/OfflineSearchEngine.kt
@@ -20,6 +20,7 @@ import com.mapbox.search.base.record.IndexableRecordResolver
 import com.mapbox.search.base.result.SearchResultFactory
 import com.mapbox.search.base.utils.AndroidKeyboardLocaleProvider
 import com.mapbox.search.base.utils.UserAgentProvider
+import com.mapbox.search.base.utils.defaultOnlineRequestTimeoutSeconds
 import com.mapbox.search.base.utils.orientation.AndroidScreenOrientationProvider
 import com.mapbox.search.common.AsyncOperationTask
 import com.mapbox.search.common.concurrent.SearchSdkMainThreadWorker
@@ -138,6 +139,93 @@ public interface OfflineSearchEngine {
         callback: OfflineSearchCallback,
     ): AsyncOperationTask = search(
         query = query,
+        options = options,
+        executor = SearchSdkMainThreadWorker.mainExecutor,
+        callback = callback,
+    )
+
+    /**
+     * Performs category search request.
+     * Each new search request cancels the previous one if it is still in progress.
+     * In this case [OfflineSearchCallback.onError] will be called with [com.mapbox.search.common.SearchCancellationException].
+     *
+     * @param categoryNames List of categories to search.
+     * @param options Category search options.
+     * @param executor Executor used for events dispatching. By default events are dispatched on the main thread.
+     * @param callback The callback to handle search result on the main thread.
+     * @return [AsyncOperationTask] object which allows to cancel the request.
+     */
+    @MapboxExperimental
+    public fun categorySearch(
+        categoryNames: List<String>,
+        options: OfflineCategorySearchOptions,
+        executor: Executor,
+        callback: OfflineSearchCallback,
+    ): AsyncOperationTask
+
+    /**
+     * Performs category search request.
+     * Each new search request cancels the previous one if it is still in progress.
+     * In this case [OfflineSearchCallback.onError] will be called with [com.mapbox.search.common.SearchCancellationException].
+     *
+     * @param categoryNames List of categories to search.
+     * @param options Category search options.
+     * @param callback The callback to handle search result. Events are dispatched on the main thread.
+     * @return [AsyncOperationTask] object which allows to cancel the request.
+     */
+    @MapboxExperimental
+    public fun categorySearch(
+        categoryNames: List<String>,
+        options: OfflineCategorySearchOptions,
+        callback: OfflineSearchCallback,
+    ): AsyncOperationTask = categorySearch(
+        categoryNames = categoryNames,
+        options = options,
+        executor = SearchSdkMainThreadWorker.mainExecutor,
+        callback = callback,
+    )
+
+    /**
+     * Performs category search request.
+     * Each new search request cancels the previous one if it is still in progress.
+     * In this case [OfflineSearchCallback.onError] will be called with [com.mapbox.search.common.SearchCancellationException].
+     *
+     * @param categoryName Name of category to search.
+     * @param options Category search options.
+     * @param executor Executor used for events dispatching. By default events are dispatched on the main thread.
+     * @param callback The callback to handle search result on the main thread.
+     * @return [AsyncOperationTask] object which allows to cancel the request.
+     */
+    @MapboxExperimental
+    public fun categorySearch(
+        categoryName: String,
+        options: OfflineCategorySearchOptions,
+        executor: Executor,
+        callback: OfflineSearchCallback,
+    ): AsyncOperationTask = categorySearch(
+        categoryNames = listOf(categoryName),
+        options = options,
+        executor = executor,
+        callback = callback,
+    )
+
+    /**
+     * Performs category search request.
+     * Each new search request cancels the previous one if it is still in progress.
+     * In this case [OfflineSearchCallback.onError] will be called with [com.mapbox.search.common.SearchCancellationException].
+     *
+     * @param categoryName Name of category to search.
+     * @param options Category search options.
+     * @param callback The callback to handle search result. Events are dispatched on the main thread.
+     * @return [AsyncOperationTask] object which allows to cancel the request.
+     */
+    @MapboxExperimental
+    public fun categorySearch(
+        categoryName: String,
+        options: OfflineCategorySearchOptions,
+        callback: OfflineSearchCallback,
+    ): AsyncOperationTask = categorySearch(
+        categoryName = categoryName,
         options = options,
         executor = SearchSdkMainThreadWorker.mainExecutor,
         callback = callback,
@@ -432,6 +520,7 @@ public interface OfflineSearchEngine {
                     apiType = CoreApiType.SBS,
                     sdkInformation = UserAgentProvider.sdkInformation(),
                     eventsUrl = null,
+                    onlineRequestTimeout = defaultOnlineRequestTimeoutSeconds(),
                 ),
                 WrapperLocationProvider(
                     LocationEngineAdapter(app, settings.locationProvider),